Rename further "free" headers
