+++ src/gtt_file_io.c
@@ -34,7 +34,7 @@
 #include "gtt_gsettings_io.h"
 #include "gtt.h"
 #include "gtt_menus.h"
-#include "plug-in.h"
+#include "gtt_plug_in.h"
 #include "gtt_preferences.h"
 #include "gtt_project.h"
 #include "proj_p.h"
