+++ src/gtt_myoaf.c
@@ -5,16 +5,13 @@
    eliminated.
 */
 
-
-
+#include "gtt_myoaf.h"
 
 #include <gtk/gtk.h>
 #if 0
 #include <liboaf/liboaf.h>
 #include <orb/orbit.h>
 #endif
-
-#include "myoaf.h"
