+++ src/gtt_status_icon.c
@@ -28,7 +28,7 @@
 
 #include <gtk/gtk.h>
 #include <gnome.h>
-#include "status-icon.h"
+#include "gtt_status_icon.h"
 #include "gtt_timer.h"
 
 extern GtkWidget *app_window;  /* global top-level window */
