+++ src/gtt_status_icon.h
